implement data segment map and simple allocator
We don't have an RW .data segment (because we always run from ROM so
far). To store RW data that does not need to be deallocated we use a
simple grows-up allocator, which is called data segment. This segment is
just that: a reserved memory area. Since there is no deallocation
involved, we don't need to worry about fragmentation.

Signed-off-by: Evgeny Iakovlev <eyakovlev3@gmail.com>